chore: refactor and format
