+++ src/collectors/indexer-collector.ts
@@ -1,145 +1,14 @@
-import axios from 'axios';
 import { CollectorOptions } from './collector';
 import { SUDTCollector } from './sudt-collector';
+import { Cell, Address, Amount, AmountUnit, SUDT } from '..';
 import {
-  Cell,
-  Address,
-  Amount,
-  AmountUnit,
-  OutPoint,
-  SUDT,
-  Script as PwScript,
-} from '..';
-
-enum ScriptType {
-  type = 'type',
-  lock = 'lock',
-}
-
-enum Order {
-  asc = 'asc',
-  desc = 'desc',
-}
-
-type HexString = string;
-type Hash = HexString;
-type HashType = 'type' | 'data';
-
-interface Script {
-  code_hash: Hash;
-  hash_type: HashType;
-  args: HexString;
-}
-
-interface SearchKey {
-  script: Script;
-  script_type: ScriptType;
-  filter?: {
-    script?: Script;
-    output_data_len_range?: [HexString, HexString];
-    output_capacity_range?: [HexString, HexString];
-    block_range?: [HexString, HexString];
-  };
-}
-
-interface IndexerCell {
-  block_number: HexString;
-  out_point: {
-    index: HexString;
-    tx_hash: Hash;
-  };
-  output: {
-    capacity: HexString;
-    lock: Script;
-    type?: Script;
-  };
-  output_data: HexString;
-  tx_index: HexString;
-}
-
-interface TerminatorResult {
-  stop: boolean;
-  push: boolean;
-}
-
-declare type Terminator = (
-  index: number,
-  cell: IndexerCell
-) => TerminatorResult;
-
-const DefaultTerminator: Terminator = (_index, _cell) => {
-  return { stop: false, push: true };
-};
-
-function IndexerCellToCell(cell: IndexerCell): Cell {
-  return new Cell(
-    new Amount(cell.output.capacity, AmountUnit.shannon),
-    PwScript.fromRPC(cell.output.lock),
-    PwScript.fromRPC(cell.output.type),
-    OutPoint.fromRPC(cell.out_point),
-    cell.output_data
-  );
-}
-
-export class CkbIndexer {
-  constructor(public ckbIndexerUrl: string) {}
-
-  async request(method: string, params?: any): Promise<any> {
-    const data = {
-      id: 0,
-      jsonrpc: '2.0',
-      method,
-      params,
-    };
-    // console.dir({data, url: this.ckbIndexerUrl}, {depth: null});
-    const res = await axios.post(this.ckbIndexerUrl, data);
-    // console.dir(res, {depth: null});
-    if (res.status !== 200) {
-      throw new Error(`indexer request failed with HTTP code ${res.status}`);
-    }
-    if (res.data.error !== undefined) {
-      throw new Error(
-        `indexer request rpc failed with error: ${JSON.stringify(
-          res.data.error
-        )}`
-      );
-    }
-    return res.data.result;
-  }
-
-  public async getCells(
-    searchKey: SearchKey,
-    terminator: Terminator = DefaultTerminator,
-    {
-      sizeLimit = 0x100,
-      order = Order.asc,
-    }: { sizeLimit?: number; order?: Order } = {}
-  ): Promise<IndexerCell[]> {
-    const infos: IndexerCell[] = [];
-    let cursor = null;
-    let index = 0;
-    const params = [searchKey, order, `0x${sizeLimit.toString(16)}`, cursor];
-    while (true) {
-      const res = await this.request('get_cells', params);
-      const liveCells = res.objects;
-      cursor = res.lastCursor;
-      for (const cell of liveCells) {
-        const { stop, push } = terminator(index, cell);
-        if (push) {
-          infos.push(cell);
-        }
-        if (stop) {
-          return infos;
-        }
-        index++;
-      }
-      if (liveCells.length < sizeLimit) {
-        break;
-      }
-    }
-    return infos;
-  }
-}
+  CkbIndexer,
+  HexString,
+  IndexerCellToCell,
+  Script,
+  ScriptType,
+  Terminator,
+} from '../helpers/ckb-indexer';
 
 export class IndexerCollector extends SUDTCollector {
   private indexer: CkbIndexer;
@@ -158,7 +27,7 @@ export class IndexerCollector extends SUDTCollector {
       },
       script_type: ScriptType.lock,
       filter: {
-        output_data_len_range: ['0x0', '0x1'],
+        output_data_len_range: ['0x0', '0x1'] as [HexString, HexString],
       },
     };
     // console.log({searchKey});
@@ -201,7 +70,7 @@ export class IndexerCollector extends SUDTCollector {
       },
       script_type: ScriptType.lock,
       filter: {
-        output_data_len_range: ['0x0', '0x1'],
+        output_data_len_range: ['0x0', '0x1'] as [HexString, HexString],
       },
     };
     const cells = await this.indexer.getCells(searchKey, terminator);
@@ -219,7 +88,7 @@ export class IndexerCollector extends SUDTCollector {
       },
       script_type: ScriptType.lock,
       filter: {
-        script: sudt.toTypeScript().serializeJson(),
+        script: sudt.toTypeScript().serializeJson() as Script,
       },
     };
     // console.log({searchKey});
@@ -250,7 +119,7 @@ export class IndexerCollector extends SUDTCollector {
       },
       script_type: ScriptType.lock,
       filter: {
-        script: sudt.toTypeScript().serializeJson(),
+        script: sudt.toTypeScript().serializeJson() as Script,
       },
     };
     let accCapacity = Amount.ZERO;
